cleaned up layout and added more high resolution short cv
The layout had the problem that at small screens or when you only you safari on parts of your desctop, then some of the elements would float too much to the left. This update solves this (hopefuly). Basically what I did is forcing the conventions for the header to the rest of the layout elements. On top of that I added a more high resolution version of the short CV and optimized the fotos therein a bit.