content straemline
I changed the order of the content so that the socials are at the end and the publications and projects are after each other. This is also makes it so that the latter two now appear as a single content block (same background and style), which I think is better. On top of that I changed the background of the socials to dark green so that the colors mirror the header/cv sequence. I find that detail quite enjoyable.